+++ experimental/Pomdog.Experimental/Graphics/LineBatch.cpp
@@ -122,12 +122,12 @@ void LineBatch::Impl::Flush()
 {
 	POMDOG_ASSERT(!vertices.empty());
 	POMDOG_ASSERT(vertices.size() <= MaxVertexCount);
-	vertexBuffer->SetData(vertices.data(), static_cast<std::uint32_t>(vertices.size()));
+	vertexBuffer->SetData(vertices.data(), vertices.size());
 
 	graphicsContext->SetVertexBuffer(vertexBuffer);
 	graphicsContext->SetEffectPass(effectPass);
 	graphicsContext->SetConstantBuffers(constantBuffers);
-	graphicsContext->Draw(PrimitiveTopology::LineList, static_cast<std::uint32_t>(vertices.size()));
+	graphicsContext->Draw(PrimitiveTopology::LineList, vertices.size());
 
 	vertices.clear();
 }
